Pulling out a comparator factory
The idea was to add a  property to the Column so that when it's , you might want to apply a constructor to the data to make it into a type which can be compared using built-in operators. The example I had to work on was the readting_ts value, which is a string, but the string represents a date. The <, >, == etc operators work on Dates. I didn't trust that '2001-01-03T00:01:03' would compare as greater than to '1999-12-10T23:23:23'. I should check it...